requirements.txt
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
diff-match-patch==20241021
Django==5.2.5
django-allauth==65.11.0
django-crum==0.7.9
django-import-export==4.3.9
google-auth==2.40.3
google-auth-oauthlib==1.2.2
idna==3.10
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
tablib==3.8.0
urllib3==2.5.0
